+++ R/RcppExports.R
@@ -73,8 +73,8 @@ libgdalcubes_create_view <- function(v) {
     .Call('_gdalcubes_libgdalcubes_create_view', PACKAGE = 'gdalcubes', v)
 }
 
-libgdalcubes_create_image_collection_cube <- function(pin, chunk_sizes, v = NULL) {
-    .Call('_gdalcubes_libgdalcubes_create_image_collection_cube', PACKAGE = 'gdalcubes', pin, chunk_sizes, v)
+libgdalcubes_create_image_collection_cube <- function(pin, chunk_sizes, mask, v = NULL) {
+    .Call('_gdalcubes_libgdalcubes_create_image_collection_cube', PACKAGE = 'gdalcubes', pin, chunk_sizes, mask, v)
 }
 
 libgdalcubes_create_dummy_cube <- function(v, nbands, fill, chunk_sizes) {
@@ -127,6 +127,10 @@ libgdalcubes_eval_cube <- function(pin, outfile, compression_level = 0L) {
 
 libgdalcubes_create_stream_cube <- function(pin, cmd) {
     .Call('_gdalcubes_libgdalcubes_create_stream_cube', PACKAGE = 'gdalcubes', pin, cmd)
+}
+
+libgdalcubes_create_fill_time_cube <- function(pin, method) {
+    .Call('_gdalcubes_libgdalcubes_create_fill_time_cube', PACKAGE = 'gdalcubes', pin, method)
 }
 
 libgdalcubes_set_threads <- function(n) {