+++ NEWS.md
@@ -1,3 +1,11 @@
+
+# gdalcubes 0.1.0.9999
+
+* rename `filter_predicate()` -> `filter_pixel()`
+* new operator `fill_time()` to fill NA pixels of data cubes based on time-series interpolation
+* add masking based on pixel band values while reading images, see `?image_mask`
+
+
 # gdalcubes 0.1.0 (2019-03-13)
 
 * First release